// front-orders/src/modules/dashboard/modules/subscription/enum/distributiveFund.enum.ts
export enum fundDistributiveFundEnum {
  'id' = 18,
  'distributiveFundName' = 'Prudential Estrategia Distributiva Global Fondo de Fondos FMIV',
  'codFund' = '006',
}
export enum SerieCodEnum {
  SERIE_A = '000',
  SERIE_B = '001',
}
export enum ScheduledRescueEnum {
  COD_FORMA_PAGO = '02',
  COD_CRONOGRAM = '1',
  COD_ADMIN = '002',
  TYPE_IDENTIFICATION = '01',
}